fix: separated the Event Listener from the History Event Handlers
